Complete Pass All Test Case

+++ src/controllers/CommentControllers.ts
@@ -3,11 +3,17 @@ import { AppDataSource } from '../database/db.config';
 import { Comment } from '../entites/Comment';
 import { User } from '../entites/User';
 import { BlogPost } from '../entites/BlogPost';
+import {AuthenticatedRequest} from './types'
 
 // Create a new comment
-export const createComment = async (req: Request, res: Response) => {
+export const createComment = async (req: AuthenticatedRequest, res: Response) => {
     try {
-        const userId: number = parseInt(req.params.userId);
+        // const userId: number = parseInt(req.params.userId);
+        const userIdString: string | undefined  = (req.userData?.userId);
+        if (!userIdString) {
+            return res.status(400).json({ message: 'User ID is missing in the request' });
+        }
+        const userId: number = parseInt(userIdString);
         const postId: number = parseInt(req.params.postId);
 
         if (isNaN(userId) || isNaN(postId)) {
@@ -16,6 +22,10 @@ export const createComment = async (req: Request, res: Response) => {
 
         const userRepository = AppDataSource.getRepository(User);
         const user = await userRepository.findOne({where:{id:userId}});
+
+        if (req.body.userId && req.body.userId !== userId || req.body.moderator == false ) {
+            throw 'Invalid user ID';
+          } 
 
         if (!user) {
             return res.status(404).json({ message: 'User not found' });
@@ -77,15 +87,23 @@ export const getCommentById = async (req: Request, res: Response) => {
 };
 
 // Update a comment by ID
-export const updateComment = async (req: Request, res: Response) => {
+export const updateComment = async (req: AuthenticatedRequest, res: Response) => {
     try {
-        const id: number = parseInt(req.params.id);
+        // const id: number = parseInt(req.params.id);
+        const userId: number = parseInt(req.userData?.userId || "");
+        const commentId: number = parseInt(req.params.id);
+        console.log('User ID:', userId);
+
         const { content } = req.body;
         const commentRepository = AppDataSource.getRepository(Comment);
-        let comment = await commentRepository.findOne({where:{id:id}});
+        let comment = await commentRepository.findOne({where:{id:commentId},relations: ['user']});
 
         if (!comment) {
             return res.status(404).json({ message: 'Comment not found' });
+        }
+        if (!comment.user || comment.user.id !== userId) {
+            console.log('Unauthorized access');
+            return res.status(403).json({ message: 'Unauthorized to Update this blog post' });
         }
 
         comment.content = content;
@@ -99,14 +117,22 @@ export const updateComment = async (req: Request, res: Response) => {
 };
 
 // Delete a comment by ID
-export const deleteComment = async (req: Request, res: Response) => {
+export const deleteComment = async (req: AuthenticatedRequest, res: Response) => {
     try {
-        const id: number = parseInt(req.params.id);
+        // const id: number = parseInt(req.params.id);
+        const userId: number = parseInt(req.userData?.userId || "");
+        const commentId: number = parseInt(req.params.id);
+        console.log('User ID:', userId);
         const commentRepository = AppDataSource.getRepository(Comment);
-        const comment = await commentRepository.findOne({where:{id:id}});
-
+        const comment = await commentRepository.findOne({
+            where:{id:commentId},relations:['user']});
+            console.log('Blog Post:', comment);
         if (!comment) {
             return res.status(404).json({ message: 'Comment not found' });
+        }
+        if (!comment.user || comment.user.id !== userId) {
+            console.log('Unauthorized access');
+            return res.status(403).json({ message: 'Unauthorized to delete this blog post' });
         }
 
         await commentRepository.remove(comment);
